add SPM support
