+++ Sources/ParallaxPagerView/ParallaxPagerView.plist
