+++ Sources/ParallaxPagerView/Tabs/TabsView.xib
